requirements.txt
elasticsearch
pandas
fastapi
uvicorn
python-dotenv
openai
pypdf
requests
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
